updates: fix the event route to take account of the userId in the controllers

+++ cmd/models/event.go
@@ -35,12 +35,12 @@ func (m *EventModel) GetOneEventById(ctx context.Context, id string)(pgx.Row){
 	return config.Db.QueryRow(ctx, sql, id)
 }
 
-func (m *EventModel)DeleteEvent(ctx context.Context, id string)(pgconn.CommandTag, error){
-	sql := "DELETE FROM events WHERE id = $1"
-	return config.Db.Exec(ctx, sql, id)	
+func (m *EventModel)DeleteEvent(ctx context.Context, id, creator_id string)(pgconn.CommandTag, error){
+	sql := "DELETE FROM events WHERE id = $1 AND creator_id = $2"
+	return config.Db.Exec(ctx, sql, id,creator_id)	
 }
 
-func (m *EventModel) UpdateEvent(ctx context.Context, name,id string)(pgx.Row){
-	sql := "UPDATE events SET name = $1 WHERE id = $2 RETURNING id, name, creator_id, created_at"
-	return config.Db.QueryRow(ctx,sql, name, id)
+func (m *EventModel) UpdateEvent(ctx context.Context, name,id,creator_id string)(pgx.Row){
+	sql := "UPDATE events SET name = $1 WHERE id = $2 AND creator_id = $3 RETURNING id, name, creator_id, created_at"
+	return config.Db.QueryRow(ctx,sql, name, id, creator_id)
 }
